Remove unneeded client subinterfaces
Since the removal of the transport client these interfaces all have
exactly one implementation. There's no need for the extra layer of
abstraction any more so this commit removes it.
