+++ server/src/main/java/org/elasticsearch/client/internal/IndicesAdminClient.java
@@ -10,9 +10,14 @@ package org.elasticsearch.client.internal;
 
 import org.elasticsearch.action.ActionFuture;
 import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.ActionRequest;
+import org.elasticsearch.action.ActionResponse;
+import org.elasticsearch.action.ActionType;
 import org.elasticsearch.action.admin.indices.alias.IndicesAliasesRequest;
 import org.elasticsearch.action.admin.indices.alias.IndicesAliasesRequestBuilder;
 import org.elasticsearch.action.admin.indices.alias.IndicesAliasesResponse;
+import org.elasticsearch.action.admin.indices.alias.TransportIndicesAliasesAction;
+import org.elasticsearch.action.admin.indices.alias.get.GetAliasesAction;
 import org.elasticsearch.action.admin.indices.alias.get.GetAliasesRequest;
 import org.elasticsearch.action.admin.indices.alias.get.GetAliasesRequestBuilder;
 import org.elasticsearch.action.admin.indices.alias.get.GetAliasesResponse;
@@ -20,64 +25,88 @@ import org.elasticsearch.action.admin.indices.analyze.AnalyzeAction;
 import org.elasticsearch.action.admin.indices.analyze.AnalyzeRequestBuilder;
 import org.elasticsearch.action.admin.indices.cache.clear.ClearIndicesCacheRequest;
 import org.elasticsearch.action.admin.indices.cache.clear.ClearIndicesCacheRequestBuilder;
+import org.elasticsearch.action.admin.indices.cache.clear.TransportClearIndicesCacheAction;
 import org.elasticsearch.action.admin.indices.close.CloseIndexRequest;
 import org.elasticsearch.action.admin.indices.close.CloseIndexRequestBuilder;
 import org.elasticsearch.action.admin.indices.close.CloseIndexResponse;
+import org.elasticsearch.action.admin.indices.close.TransportCloseIndexAction;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequestBuilder;
 import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
+import org.elasticsearch.action.admin.indices.create.TransportCreateIndexAction;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequestBuilder;
+import org.elasticsearch.action.admin.indices.delete.TransportDeleteIndexAction;
+import org.elasticsearch.action.admin.indices.flush.FlushAction;
 import org.elasticsearch.action.admin.indices.flush.FlushRequest;
 import org.elasticsearch.action.admin.indices.flush.FlushRequestBuilder;
+import org.elasticsearch.action.admin.indices.forcemerge.ForceMergeAction;
 import org.elasticsearch.action.admin.indices.forcemerge.ForceMergeRequest;
 import org.elasticsearch.action.admin.indices.forcemerge.ForceMergeRequestBuilder;
+import org.elasticsearch.action.admin.indices.get.GetIndexAction;
 import org.elasticsearch.action.admin.indices.get.GetIndexRequest;
 import org.elasticsearch.action.admin.indices.get.GetIndexRequestBuilder;
 import org.elasticsearch.action.admin.indices.get.GetIndexResponse;
+import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsAction;
 import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsRequest;
 import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsRequestBuilder;
 import org.elasticsearch.action.admin.indices.mapping.get.GetFieldMappingsResponse;
+import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsAction;
 import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsRequest;
 import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsRequestBuilder;
 import org.elasticsearch.action.admin.indices.mapping.get.GetMappingsResponse;
 import org.elasticsearch.action.admin.indices.mapping.put.PutMappingRequest;
 import org.elasticsearch.action.admin.indices.mapping.put.PutMappingRequestBuilder;
+import org.elasticsearch.action.admin.indices.mapping.put.TransportPutMappingAction;
+import org.elasticsearch.action.admin.indices.open.OpenIndexAction;
 import org.elasticsearch.action.admin.indices.open.OpenIndexRequest;
 import org.elasticsearch.action.admin.indices.open.OpenIndexRequestBuilder;
 import org.elasticsearch.action.admin.indices.open.OpenIndexResponse;
 import org.elasticsearch.action.admin.indices.readonly.AddIndexBlockRequest;
 import org.elasticsearch.action.admin.indices.readonly.AddIndexBlockRequestBuilder;
 import org.elasticsearch.action.admin.indices.readonly.AddIndexBlockResponse;
+import org.elasticsearch.action.admin.indices.readonly.TransportAddIndexBlockAction;
+import org.elasticsearch.action.admin.indices.recovery.RecoveryAction;
 import org.elasticsearch.action.admin.indices.recovery.RecoveryRequest;
 import org.elasticsearch.action.admin.indices.recovery.RecoveryRequestBuilder;
 import org.elasticsearch.action.admin.indices.recovery.RecoveryResponse;
+import org.elasticsearch.action.admin.indices.refresh.RefreshAction;
 import org.elasticsearch.action.admin.indices.refresh.RefreshRequest;
 import org.elasticsearch.action.admin.indices.refresh.RefreshRequestBuilder;
 import org.elasticsearch.action.admin.indices.resolve.ResolveIndexAction;
+import org.elasticsearch.action.admin.indices.rollover.RolloverAction;
 import org.elasticsearch.action.admin.indices.rollover.RolloverRequest;
 import org.elasticsearch.action.admin.indices.rollover.RolloverRequestBuilder;
 import org.elasticsearch.action.admin.indices.rollover.RolloverResponse;
 import org.elasticsearch.action.admin.indices.segments.IndicesSegmentResponse;
+import org.elasticsearch.action.admin.indices.segments.IndicesSegmentsAction;
 import org.elasticsearch.action.admin.indices.segments.IndicesSegmentsRequest;
 import org.elasticsearch.action.admin.indices.segments.IndicesSegmentsRequestBuilder;
+import org.elasticsearch.action.admin.indices.settings.get.GetSettingsAction;
 import org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequest;
 import org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequestBuilder;
 import org.elasticsearch.action.admin.indices.settings.get.GetSettingsResponse;
+import org.elasticsearch.action.admin.indices.settings.put.TransportUpdateSettingsAction;
 import org.elasticsearch.action.admin.indices.settings.put.UpdateSettingsRequest;
 import org.elasticsearch.action.admin.indices.settings.put.UpdateSettingsRequestBuilder;
+import org.elasticsearch.action.admin.indices.shrink.ResizeAction;
 import org.elasticsearch.action.admin.indices.shrink.ResizeRequest;
 import org.elasticsearch.action.admin.indices.shrink.ResizeRequestBuilder;
+import org.elasticsearch.action.admin.indices.stats.IndicesStatsAction;
 import org.elasticsearch.action.admin.indices.stats.IndicesStatsRequest;
 import org.elasticsearch.action.admin.indices.stats.IndicesStatsRequestBuilder;
 import org.elasticsearch.action.admin.indices.stats.IndicesStatsResponse;
 import org.elasticsearch.action.admin.indices.template.delete.DeleteIndexTemplateRequest;
 import org.elasticsearch.action.admin.indices.template.delete.DeleteIndexTemplateRequestBuilder;
+import org.elasticsearch.action.admin.indices.template.delete.TransportDeleteIndexTemplateAction;
+import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesAction;
 import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesRequest;
 import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesRequestBuilder;
 import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesResponse;
 import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateRequest;
 import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateRequestBuilder;
+import org.elasticsearch.action.admin.indices.template.put.TransportPutIndexTemplateAction;
+import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryAction;
 import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryRequest;
 import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryRequestBuilder;
 import org.elasticsearch.action.admin.indices.validate.query.ValidateQueryResponse;
@@ -85,533 +114,364 @@ import org.elasticsearch.action.support.broadcast.BroadcastResponse;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.cluster.metadata.IndexMetadata.APIBlock;
 import org.elasticsearch.core.Nullable;
+import org.elasticsearch.threadpool.ThreadPool;
 
 /**
  * Administrative actions/operations against indices.
  *
  * @see AdminClient#indices()
  */
-public interface IndicesAdminClient extends ElasticsearchClient {
-
-    /**
-     * Indices stats.
-     */
-    ActionFuture<IndicesStatsResponse> stats(IndicesStatsRequest request);
-
-    /**
-     * Indices stats.
-     */
-    void stats(IndicesStatsRequest request, ActionListener<IndicesStatsResponse> listener);
-
-    /**
-     * Indices stats.
-     */
-    IndicesStatsRequestBuilder prepareStats(String... indices);
-
-    /**
-     * Indices recoveries
-     */
-    ActionFuture<RecoveryResponse> recoveries(RecoveryRequest request);
-
-    /**
-     *Indices recoveries
-     */
-    void recoveries(RecoveryRequest request, ActionListener<RecoveryResponse> listener);
-
-    /**
-     * Indices recoveries
-     */
-    RecoveryRequestBuilder prepareRecoveries(String... indices);
-
-    /**
-     * The segments of one or more indices.
-     *
-     * @param request The indices segments request
-     * @return The result future
-     */
-    ActionFuture<IndicesSegmentResponse> segments(IndicesSegmentsRequest request);
-
-    /**
-     * The segments of one or more indices.
-     *
-     * @param request  The indices segments request
-     * @param listener A listener to be notified with a result
-     */
-    void segments(IndicesSegmentsRequest request, ActionListener<IndicesSegmentResponse> listener);
-
-    /**
-     * The segments of one or more indices.
-     */
-    IndicesSegmentsRequestBuilder prepareSegments(String... indices);
-
-    /**
-     * Creates an index using an explicit request allowing to specify the settings of the index.
-     *
-     * @param request The create index request
-     * @return The result future
-     */
-    ActionFuture<CreateIndexResponse> create(CreateIndexRequest request);
-
-    /**
-     * Creates an index using an explicit request allowing to specify the settings of the index.
-     *
-     * @param request  The create index request
-     * @param listener A listener to be notified with a result
-     */
-    void create(CreateIndexRequest request, ActionListener<CreateIndexResponse> listener);
-
-    /**
-     * Creates an index using an explicit request allowing to specify the settings of the index.
-     *
-     * @param index The index name to create
-     */
-    CreateIndexRequestBuilder prepareCreate(String index);
-
-    /**
-     * Deletes an index based on the index name.
-     *
-     * @param request The delete index request
-     * @return The result future
-     */
-    ActionFuture<AcknowledgedResponse> delete(DeleteIndexRequest request);
-
-    /**
-     * Deletes an index based on the index name.
-     *
-     * @param request  The delete index request
-     * @param listener A listener to be notified with a result
-     */
-    void delete(DeleteIndexRequest request, ActionListener<AcknowledgedResponse> listener);
-
-    /**
-     * Deletes an index based on the index name.
-     *
-     * @param indices The indices to delete. Use "_all" to delete all indices.
-     */
-    DeleteIndexRequestBuilder prepareDelete(String... indices);
-
-    /**
-     * Closes an index based on the index name.
-     *
-     * @param request The close index request
-     * @return The result future
-     */
-    ActionFuture<CloseIndexResponse> close(CloseIndexRequest request);
-
-    /**
-     * Closes an index based on the index name.
-     *
-     * @param request  The close index request
-     * @param listener A listener to be notified with a result
-     */
-    void close(CloseIndexRequest request, ActionListener<CloseIndexResponse> listener);
-
-    /**
-     * Closes one or more indices based on their index name.
-     *
-     * @param indices The name of the indices to close
-     */
-    CloseIndexRequestBuilder prepareClose(String... indices);
-
-    /**
-     * Open an index based on the index name.
-     *
-     * @param request The open index request
-     * @return The result future
-     */
-    ActionFuture<OpenIndexResponse> open(OpenIndexRequest request);
-
-    /**
-     * Open an index based on the index name.
-     *
-     * @param request  The open index request
-     * @param listener A listener to be notified with a result
-     */
-    void open(OpenIndexRequest request, ActionListener<OpenIndexResponse> listener);
-
-    /**
-     * Adds a block to an index
-     *
-     * @param block   The block to add
-     * @param indices The name of the indices to add the block to
-     */
-    AddIndexBlockRequestBuilder prepareAddBlock(APIBlock block, String... indices);
-
-    /**
-     * Adds a block to an index
-     *
-     * @param request  The add index block request
-     * @param listener A listener to be notified with a result
-     */
-    void addBlock(AddIndexBlockRequest request, ActionListener<AddIndexBlockResponse> listener);
-
-    /**
-     * Opens one or more indices based on their index name.
-     *
-     * @param indices The name of the indices to open
-     */
-    OpenIndexRequestBuilder prepareOpen(String... indices);
-
-    /**
-     * Explicitly refresh one or more indices (making the content indexed since the last refresh searchable).
-     *
-     * @param request The refresh request
-     * @return The result future
-     */
-    ActionFuture<BroadcastResponse> refresh(RefreshRequest request);
-
-    /**
-     * Explicitly refresh one or more indices (making the content indexed since the last refresh searchable).
-     *
-     * @param request  The refresh request
-     * @param listener A listener to be notified with a result
-     */
-    void refresh(RefreshRequest request, ActionListener<BroadcastResponse> listener);
-
-    /**
-     * Explicitly refresh one or more indices (making the content indexed since the last refresh searchable).
-     */
-    RefreshRequestBuilder prepareRefresh(String... indices);
-
-    /**
-     * Explicitly flush one or more indices (releasing memory from the node).
-     *
-     * @param request The flush request
-     * @return A result future
-     */
-    ActionFuture<BroadcastResponse> flush(FlushRequest request);
-
-    /**
-     * Explicitly flush one or more indices (releasing memory from the node).
-     *
-     * @param request  The flush request
-     * @param listener A listener to be notified with a result
-     */
-    void flush(FlushRequest request, ActionListener<BroadcastResponse> listener);
-
-    /**
-     * Explicitly flush one or more indices (releasing memory from the node).
-     */
-    FlushRequestBuilder prepareFlush(String... indices);
-
-    /**
-     * Explicitly force merge one or more indices into a the number of segments.
-     *
-     * @param request The optimize request
-     * @return A result future
-     */
-    ActionFuture<BroadcastResponse> forceMerge(ForceMergeRequest request);
-
-    /**
-     * Explicitly force merge one or more indices into a the number of segments.
-     *
-     * @param request  The force merge request
-     * @param listener A listener to be notified with a result
-     */
-    void forceMerge(ForceMergeRequest request, ActionListener<BroadcastResponse> listener);
-
-    /**
-     * Explicitly force merge one or more indices into a the number of segments.
-     */
-    ForceMergeRequestBuilder prepareForceMerge(String... indices);
-
-    /**
-     * Get the complete mappings of one or more types
-     */
-    void getMappings(GetMappingsRequest request, ActionListener<GetMappingsResponse> listener);
-
-    /**
-     * Get the complete mappings of one or more types
-     */
-    ActionFuture<GetMappingsResponse> getMappings(GetMappingsRequest request);
-
-    /**
-     * Get the complete mappings of one or more types
-     */
-    GetMappingsRequestBuilder prepareGetMappings(String... indices);
-
-    /**
-     * Get the mappings of specific fields
-     */
-    void getFieldMappings(GetFieldMappingsRequest request, ActionListener<GetFieldMappingsResponse> listener);
-
-    /**
-     * Get the mappings of specific fields
-     */
-    GetFieldMappingsRequestBuilder prepareGetFieldMappings(String... indices);
-
-    /**
-     * Get the mappings of specific fields
-     */
-    ActionFuture<GetFieldMappingsResponse> getFieldMappings(GetFieldMappingsRequest request);
-
-    /**
-     * Add mapping definition for a type into one or more indices.
-     *
-     * @param request The create mapping request
-     * @return A result future
-     */
-    ActionFuture<AcknowledgedResponse> putMapping(PutMappingRequest request);
-
-    /**
-     * Add mapping definition for a type into one or more indices.
-     *
-     * @param request  The create mapping request
-     * @param listener A listener to be notified with a result
-     */
-    void putMapping(PutMappingRequest request, ActionListener<AcknowledgedResponse> listener);
-
-    /**
-     * Add mapping definition for a type into one or more indices.
-     */
-    PutMappingRequestBuilder preparePutMapping(String... indices);
-
-    /**
-     * Allows to add/remove aliases from indices.
-     *
-     * @param request The index aliases request
-     * @return The result future
-     */
-    ActionFuture<IndicesAliasesResponse> aliases(IndicesAliasesRequest request);
-
-    /**
-     * Allows to add/remove aliases from indices.
-     *
-     * @param request  The index aliases request
-     * @param listener A listener to be notified with a result
-     */
-    void aliases(IndicesAliasesRequest request, ActionListener<IndicesAliasesResponse> listener);
-
-    /**
-     * Allows to add/remove aliases from indices.
-     */
-    IndicesAliasesRequestBuilder prepareAliases();
-
-    /**
-     * Get specific index aliases that exists in particular indices and / or by name.
-     *
-     * @param request The result future
-     */
-    ActionFuture<GetAliasesResponse> getAliases(GetAliasesRequest request);
-
-    /**
-     * Get specific index aliases that exists in particular indices and / or by name.
-     *
-     * @param request  The index aliases request
-     * @param listener A listener to be notified with a result
-     */
-    void getAliases(GetAliasesRequest request, ActionListener<GetAliasesResponse> listener);
-
-    /**
-     * Get specific index aliases that exists in particular indices and / or by name.
-     */
-    GetAliasesRequestBuilder prepareGetAliases(String... aliases);
-
-    /**
-     * Get index metadata for particular indices.
-     *
-     * @param request The result future
-     */
-    ActionFuture<GetIndexResponse> getIndex(GetIndexRequest request);
-
-    /**
-     * Get index metadata for particular indices.
-     *
-     * @param request  The index aliases request
-     * @param listener A listener to be notified with a result
-     */
-    void getIndex(GetIndexRequest request, ActionListener<GetIndexResponse> listener);
-
-    /**
-     * Get index metadata for particular indices.
-     */
-    GetIndexRequestBuilder prepareGetIndex();
-
-    /**
-     * Clear indices cache.
-     *
-     * @param request The clear indices cache request
-     * @return The result future
-     */
-    ActionFuture<BroadcastResponse> clearCache(ClearIndicesCacheRequest request);
-
-    /**
-     * Clear indices cache.
-     *
-     * @param request  The clear indices cache request
-     * @param listener A listener to be notified with a result
-     */
-    void clearCache(ClearIndicesCacheRequest request, ActionListener<BroadcastResponse> listener);
-
-    /**
-     * Clear indices cache.
-     */
-    ClearIndicesCacheRequestBuilder prepareClearCache(String... indices);
-
-    /**
-     * Updates settings of one or more indices.
-     *
-     * @param request the update settings request
-     * @return The result future
-     */
-    ActionFuture<AcknowledgedResponse> updateSettings(UpdateSettingsRequest request);
-
-    /**
-     * Updates settings of one or more indices.
-     *
-     * @param request  the update settings request
-     * @param listener A listener to be notified with the response
-     */
-    void updateSettings(UpdateSettingsRequest request, ActionListener<AcknowledgedResponse> listener);
-
-    /**
-     * Update indices settings.
-     */
-    UpdateSettingsRequestBuilder prepareUpdateSettings(String... indices);
-
-    /**
-     * Analyze text under the provided index.
-     */
-    ActionFuture<AnalyzeAction.Response> analyze(AnalyzeAction.Request request);
-
-    /**
-     * Analyze text under the provided index.
-     */
-    void analyze(AnalyzeAction.Request request, ActionListener<AnalyzeAction.Response> listener);
-
-    /**
-     * Analyze text under the provided index.
-     *
-     * @param index The index name
-     * @param text  The text to analyze
-     */
-    AnalyzeRequestBuilder prepareAnalyze(@Nullable String index, String text);
-
-    /**
-     * Analyze text.
-     *
-     * @param text The text to analyze
-     */
-    AnalyzeRequestBuilder prepareAnalyze(String text);
-
-    /**
-     * Analyze text/texts.
-     *
-     */
-    AnalyzeRequestBuilder prepareAnalyze();
-
-    /**
-     * Puts an index template.
-     */
-    ActionFuture<AcknowledgedResponse> putTemplate(PutIndexTemplateRequest request);
-
-    /**
-     * Puts an index template.
-     */
-    void putTemplate(PutIndexTemplateRequest request, ActionListener<AcknowledgedResponse> listener);
-
-    /**
-     * Puts an index template.
-     *
-     * @param name The name of the template.
-     */
-    PutIndexTemplateRequestBuilder preparePutTemplate(String name);
-
-    /**
-     * Deletes an index template.
-     */
-    void deleteTemplate(DeleteIndexTemplateRequest request, ActionListener<AcknowledgedResponse> listener);
-
-    /**
-     * Deletes an index template.
-     *
-     * @param name The name of the template.
-     */
-    DeleteIndexTemplateRequestBuilder prepareDeleteTemplate(String name);
-
-    /**
-     * Gets an index template.
-     */
-    void getTemplates(GetIndexTemplatesRequest request, ActionListener<GetIndexTemplatesResponse> listener);
-
-    /**
-     * Gets an index template (optional).
-     */
-    GetIndexTemplatesRequestBuilder prepareGetTemplates(String... name);
-
-    /**
-     * Validate a query for correctness.
-     *
-     * @param request The count request
-     * @return The result future
-     */
-    ActionFuture<ValidateQueryResponse> validateQuery(ValidateQueryRequest request);
-
-    /**
-     * Validate a query for correctness.
-     *
-     * @param request  The count request
-     * @param listener A listener to be notified of the result
-     */
-    void validateQuery(ValidateQueryRequest request, ActionListener<ValidateQueryResponse> listener);
-
-    /**
-     * Validate a query for correctness.
-     */
-    ValidateQueryRequestBuilder prepareValidateQuery(String... indices);
-
-    /**
-     * Executed a per index settings get request and returns the settings for the indices specified.
-     * Note: this is a per index request and will not include settings that are set on the cluster
-     * level. This request is not exhaustive, it will not return default values for setting.
-     */
-    void getSettings(GetSettingsRequest request, ActionListener<GetSettingsResponse> listener);
-
-    /**
-     * Executed a per index settings get request.
-     * @see #getSettings(org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequest)
-     */
-    ActionFuture<GetSettingsResponse> getSettings(GetSettingsRequest request);
-
-    /**
-     * Returns a builder for a per index settings get request.
-     * @param indices the indices to fetch the setting for.
-     * @see #getSettings(org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequest)
-     */
-    GetSettingsRequestBuilder prepareGetSettings(String... indices);
-
-    /**
-     * Resize an index using an explicit request allowing to specify the settings, mappings and aliases of the target index of the index.
-     */
-    ResizeRequestBuilder prepareResizeIndex(String sourceIndex, String targetIndex);
-
-    /**
-     * Shrinks an index using an explicit request allowing to specify the settings, mappings and aliases of the target index of the index.
-     */
-    void resizeIndex(ResizeRequest request, ActionListener<CreateIndexResponse> listener);
-
-    /**
-     * Swaps the index pointed to by an alias given all provided conditions are satisfied
-     */
-    RolloverRequestBuilder prepareRolloverIndex(String sourceAlias);
-
-    /**
-     * Swaps the index pointed to by an alias given all provided conditions are satisfied
-     */
-    ActionFuture<RolloverResponse> rolloverIndex(RolloverRequest request);
-
-    /**
-     * Swaps the index pointed to by an alias given all provided conditions are satisfied
-     */
-    void rolloverIndex(RolloverRequest request, ActionListener<RolloverResponse> listener);
-
-    /**
-     * Resolves names and wildcard expressions to indices, aliases, and data streams
-     */
-    void resolveIndex(ResolveIndexAction.Request request, ActionListener<ResolveIndexAction.Response> listener);
-
-    /**
-     * Resolves names and wildcard expressions to indices, aliases, and data streams
-     */
-    ActionFuture<ResolveIndexAction.Response> resolveIndex(ResolveIndexAction.Request request);
+public class IndicesAdminClient implements ElasticsearchClient {
+
+    protected final ElasticsearchClient client;
+
+    public IndicesAdminClient(ElasticsearchClient client) {
+        this.client = client;
+    }
+
+    public ActionFuture<IndicesStatsResponse> stats(final IndicesStatsRequest request) {
+        return execute(IndicesStatsAction.INSTANCE, request);
+    }
+
+    public void stats(final IndicesStatsRequest request, final ActionListener<IndicesStatsResponse> listener) {
+        execute(IndicesStatsAction.INSTANCE, request, listener);
+    }
+
+    public IndicesStatsRequestBuilder prepareStats(String... indices) {
+        return new IndicesStatsRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<RecoveryResponse> recoveries(final RecoveryRequest request) {
+        return execute(RecoveryAction.INSTANCE, request);
+    }
+
+    public void recoveries(final RecoveryRequest request, final ActionListener<RecoveryResponse> listener) {
+        execute(RecoveryAction.INSTANCE, request, listener);
+    }
+
+    public RecoveryRequestBuilder prepareRecoveries(String... indices) {
+        return new RecoveryRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<IndicesSegmentResponse> segments(final IndicesSegmentsRequest request) {
+        return execute(IndicesSegmentsAction.INSTANCE, request);
+    }
+
+    public void segments(final IndicesSegmentsRequest request, final ActionListener<IndicesSegmentResponse> listener) {
+        execute(IndicesSegmentsAction.INSTANCE, request, listener);
+    }
+
+    public IndicesSegmentsRequestBuilder prepareSegments(String... indices) {
+        return new IndicesSegmentsRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<CreateIndexResponse> create(final CreateIndexRequest request) {
+        return execute(TransportCreateIndexAction.TYPE, request);
+    }
+
+    public void create(final CreateIndexRequest request, final ActionListener<CreateIndexResponse> listener) {
+        execute(TransportCreateIndexAction.TYPE, request, listener);
+    }
+
+    public CreateIndexRequestBuilder prepareCreate(String index) {
+        return new CreateIndexRequestBuilder(this, index);
+    }
+
+    public ActionFuture<AcknowledgedResponse> delete(final DeleteIndexRequest request) {
+        return execute(TransportDeleteIndexAction.TYPE, request);
+    }
+
+    public void delete(final DeleteIndexRequest request, final ActionListener<AcknowledgedResponse> listener) {
+        execute(TransportDeleteIndexAction.TYPE, request, listener);
+    }
+
+    public DeleteIndexRequestBuilder prepareDelete(String... indices) {
+        return new DeleteIndexRequestBuilder(this, indices);
+    }
+
+    public ActionFuture<CloseIndexResponse> close(final CloseIndexRequest request) {
+        return execute(TransportCloseIndexAction.TYPE, request);
+    }
+
+    public void close(final CloseIndexRequest request, final ActionListener<CloseIndexResponse> listener) {
+        execute(TransportCloseIndexAction.TYPE, request, listener);
+    }
+
+    public CloseIndexRequestBuilder prepareClose(String... indices) {
+        return new CloseIndexRequestBuilder(this, indices);
+    }
+
+    public ActionFuture<OpenIndexResponse> open(final OpenIndexRequest request) {
+        return execute(OpenIndexAction.INSTANCE, request);
+    }
+
+    public void open(final OpenIndexRequest request, final ActionListener<OpenIndexResponse> listener) {
+        execute(OpenIndexAction.INSTANCE, request, listener);
+    }
+
+    public AddIndexBlockRequestBuilder prepareAddBlock(APIBlock block, String... indices) {
+        return new AddIndexBlockRequestBuilder(this, block, indices);
+    }
+
+    public void addBlock(AddIndexBlockRequest request, ActionListener<AddIndexBlockResponse> listener) {
+        execute(TransportAddIndexBlockAction.TYPE, request, listener);
+    }
+
+    public OpenIndexRequestBuilder prepareOpen(String... indices) {
+        return new OpenIndexRequestBuilder(this, indices);
+    }
+
+    public ActionFuture<BroadcastResponse> refresh(final RefreshRequest request) {
+        return execute(RefreshAction.INSTANCE, request);
+    }
+
+    public void refresh(final RefreshRequest request, final ActionListener<BroadcastResponse> listener) {
+        execute(RefreshAction.INSTANCE, request, listener);
+    }
+
+    public RefreshRequestBuilder prepareRefresh(String... indices) {
+        return new RefreshRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<BroadcastResponse> flush(final FlushRequest request) {
+        return execute(FlushAction.INSTANCE, request);
+    }
+
+    public void flush(final FlushRequest request, final ActionListener<BroadcastResponse> listener) {
+        execute(FlushAction.INSTANCE, request, listener);
+    }
+
+    public FlushRequestBuilder prepareFlush(String... indices) {
+        return new FlushRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<BroadcastResponse> forceMerge(final ForceMergeRequest request) {
+        return execute(ForceMergeAction.INSTANCE, request);
+    }
+
+    public void forceMerge(final ForceMergeRequest request, final ActionListener<BroadcastResponse> listener) {
+        execute(ForceMergeAction.INSTANCE, request, listener);
+    }
+
+    public ForceMergeRequestBuilder prepareForceMerge(String... indices) {
+        return new ForceMergeRequestBuilder(this).setIndices(indices);
+    }
+
+    public void getMappings(GetMappingsRequest request, ActionListener<GetMappingsResponse> listener) {
+        execute(GetMappingsAction.INSTANCE, request, listener);
+    }
+
+    public ActionFuture<GetMappingsResponse> getMappings(GetMappingsRequest request) {
+        return execute(GetMappingsAction.INSTANCE, request);
+    }
+
+    public GetMappingsRequestBuilder prepareGetMappings(String... indices) {
+        return new GetMappingsRequestBuilder(this, indices);
+    }
+
+    public void getFieldMappings(GetFieldMappingsRequest request, ActionListener<GetFieldMappingsResponse> listener) {
+        execute(GetFieldMappingsAction.INSTANCE, request, listener);
+    }
+
+    public GetFieldMappingsRequestBuilder prepareGetFieldMappings(String... indices) {
+        return new GetFieldMappingsRequestBuilder(this, indices);
+    }
+
+    public ActionFuture<GetFieldMappingsResponse> getFieldMappings(GetFieldMappingsRequest request) {
+        return execute(GetFieldMappingsAction.INSTANCE, request);
+    }
+
+    public ActionFuture<AcknowledgedResponse> putMapping(final PutMappingRequest request) {
+        return execute(TransportPutMappingAction.TYPE, request);
+    }
+
+    public void putMapping(final PutMappingRequest request, final ActionListener<AcknowledgedResponse> listener) {
+        execute(TransportPutMappingAction.TYPE, request, listener);
+    }
+
+    public PutMappingRequestBuilder preparePutMapping(String... indices) {
+        return new PutMappingRequestBuilder(this).setIndices(indices);
+    }
+
+    @Override
+    public <Request extends ActionRequest, Response extends ActionResponse> ActionFuture<Response> execute(
+        ActionType<Response> action,
+        Request request
+    ) {
+        return client.execute(action, request);
+    }
+
+    @Override
+    public <Request extends ActionRequest, Response extends ActionResponse> void execute(
+        ActionType<Response> action,
+        Request request,
+        ActionListener<Response> listener
+    ) {
+        client.execute(action, request, listener);
+    }
+
+    @Override
+    public ThreadPool threadPool() {
+        return client.threadPool();
+    }
+
+    public ActionFuture<IndicesAliasesResponse> aliases(final IndicesAliasesRequest request) {
+        return execute(TransportIndicesAliasesAction.TYPE, request);
+    }
+
+    public void aliases(final IndicesAliasesRequest request, final ActionListener<IndicesAliasesResponse> listener) {
+        execute(TransportIndicesAliasesAction.TYPE, request, listener);
+    }
+
+    public IndicesAliasesRequestBuilder prepareAliases() {
+        return new IndicesAliasesRequestBuilder(this);
+    }
+
+    public ActionFuture<GetAliasesResponse> getAliases(GetAliasesRequest request) {
+        return execute(GetAliasesAction.INSTANCE, request);
+    }
+
+    public void getAliases(GetAliasesRequest request, ActionListener<GetAliasesResponse> listener) {
+        execute(GetAliasesAction.INSTANCE, request, listener);
+    }
+
+    public GetAliasesRequestBuilder prepareGetAliases(String... aliases) {
+        return new GetAliasesRequestBuilder(this, aliases);
+    }
+
+    public ActionFuture<GetIndexResponse> getIndex(GetIndexRequest request) {
+        return execute(GetIndexAction.INSTANCE, request);
+    }
+
+    public void getIndex(GetIndexRequest request, ActionListener<GetIndexResponse> listener) {
+        execute(GetIndexAction.INSTANCE, request, listener);
+    }
+
+    public GetIndexRequestBuilder prepareGetIndex() {
+        return new GetIndexRequestBuilder(this);
+    }
+
+    public ActionFuture<BroadcastResponse> clearCache(final ClearIndicesCacheRequest request) {
+        return execute(TransportClearIndicesCacheAction.TYPE, request);
+    }
+
+    public void clearCache(final ClearIndicesCacheRequest request, final ActionListener<BroadcastResponse> listener) {
+        execute(TransportClearIndicesCacheAction.TYPE, request, listener);
+    }
+
+    public ClearIndicesCacheRequestBuilder prepareClearCache(String... indices) {
+        return new ClearIndicesCacheRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<AcknowledgedResponse> updateSettings(final UpdateSettingsRequest request) {
+        return execute(TransportUpdateSettingsAction.TYPE, request);
+    }
+
+    public void updateSettings(final UpdateSettingsRequest request, final ActionListener<AcknowledgedResponse> listener) {
+        execute(TransportUpdateSettingsAction.TYPE, request, listener);
+    }
+
+    public UpdateSettingsRequestBuilder prepareUpdateSettings(String... indices) {
+        return new UpdateSettingsRequestBuilder(this).setIndices(indices);
+    }
+
+    public ActionFuture<AnalyzeAction.Response> analyze(final AnalyzeAction.Request request) {
+        return execute(AnalyzeAction.INSTANCE, request);
+    }
+
+    public void analyze(final AnalyzeAction.Request request, final ActionListener<AnalyzeAction.Response> listener) {
+        execute(AnalyzeAction.INSTANCE, request, listener);
+    }
+
+    public AnalyzeRequestBuilder prepareAnalyze(@Nullable String index, String text) {
+        return new AnalyzeRequestBuilder(this, index, text);
+    }
+
+    public AnalyzeRequestBuilder prepareAnalyze(String text) {
+        return new AnalyzeRequestBuilder(this, null, text);
+    }
+
+    public AnalyzeRequestBuilder prepareAnalyze() {
+        return new AnalyzeRequestBuilder(this);
+    }
+
+    public ActionFuture<AcknowledgedResponse> putTemplate(final PutIndexTemplateRequest request) {
+        return execute(TransportPutIndexTemplateAction.TYPE, request);
+    }
+
+    public void putTemplate(final PutIndexTemplateRequest request, final ActionListener<AcknowledgedResponse> listener) {
+        execute(TransportPutIndexTemplateAction.TYPE, request, listener);
+    }
+
+    public PutIndexTemplateRequestBuilder preparePutTemplate(String name) {
+        return new PutIndexTemplateRequestBuilder(this, name);
+    }
+
+    public void deleteTemplate(final DeleteIndexTemplateRequest request, final ActionListener<AcknowledgedResponse> listener) {
+        execute(TransportDeleteIndexTemplateAction.TYPE, request, listener);
+    }
+
+    public DeleteIndexTemplateRequestBuilder prepareDeleteTemplate(String name) {
+        return new DeleteIndexTemplateRequestBuilder(this, name);
+    }
+
+    public void getTemplates(final GetIndexTemplatesRequest request, final ActionListener<GetIndexTemplatesResponse> listener) {
+        execute(GetIndexTemplatesAction.INSTANCE, request, listener);
+    }
+
+    public GetIndexTemplatesRequestBuilder prepareGetTemplates(String... names) {
+        return new GetIndexTemplatesRequestBuilder(this, names);
+    }
+
+    public ActionFuture<ValidateQueryResponse> validateQuery(final ValidateQueryRequest request) {
+        return execute(ValidateQueryAction.INSTANCE, request);
+    }
+
+    public void validateQuery(final ValidateQueryRequest request, final ActionListener<ValidateQueryResponse> listener) {
+        execute(ValidateQueryAction.INSTANCE, request, listener);
+    }
+
+    public ValidateQueryRequestBuilder prepareValidateQuery(String... indices) {
+        return new ValidateQueryRequestBuilder(this).setIndices(indices);
+    }
+
+    public void getSettings(GetSettingsRequest request, ActionListener<GetSettingsResponse> listener) {
+        execute(GetSettingsAction.INSTANCE, request, listener);
+    }
+
+    public ActionFuture<GetSettingsResponse> getSettings(GetSettingsRequest request) {
+        return execute(GetSettingsAction.INSTANCE, request);
+    }
+
+    public GetSettingsRequestBuilder prepareGetSettings(String... indices) {
+        return new GetSettingsRequestBuilder(this, indices);
+    }
+
+    public ResizeRequestBuilder prepareResizeIndex(String sourceIndex, String targetIndex) {
+        return new ResizeRequestBuilder(this).setSourceIndex(sourceIndex).setTargetIndex(new CreateIndexRequest(targetIndex));
+    }
+
+    public void resizeIndex(ResizeRequest request, ActionListener<CreateIndexResponse> listener) {
+        execute(ResizeAction.INSTANCE, request, listener);
+    }
+
+    public RolloverRequestBuilder prepareRolloverIndex(String alias) {
+        return new RolloverRequestBuilder(this).setRolloverTarget(alias);
+    }
+
+    public ActionFuture<RolloverResponse> rolloverIndex(RolloverRequest request) {
+        return execute(RolloverAction.INSTANCE, request);
+    }
+
+    public void rolloverIndex(RolloverRequest request, ActionListener<RolloverResponse> listener) {
+        execute(RolloverAction.INSTANCE, request, listener);
+    }
+
+    public void resolveIndex(ResolveIndexAction.Request request, ActionListener<ResolveIndexAction.Response> listener) {
+        execute(ResolveIndexAction.INSTANCE, request, listener);
+    }
+
+    public ActionFuture<ResolveIndexAction.Response> resolveIndex(ResolveIndexAction.Request request) {
+        return execute(ResolveIndexAction.INSTANCE, request);
+    }
 }
